+++ scandir/src/scandir.rs
@@ -227,7 +227,7 @@ impl Scandir {
             options: Options {
                 root_path: check_and_expand_path(root_path)?,
                 sorted: false,
-                skip_hidden: true,
+                skip_hidden: false,
                 max_depth: usize::MAX,
                 max_file_cnt: usize::MAX,
                 dir_include: None,