+++ scandir/tests/scandir.rs
@@ -12,7 +12,7 @@ fn test_scandir() -> Result<(), Error> {
     let temp_dir = common::create_temp_file_tree(3, 3, 4, 5)?;
     let entries = Scandir::new(temp_dir.path(), Some(true))?.collect()?;
     #[cfg(unix)]
-    assert_eq!(192, entries.results.len());
+    assert_eq!(210, entries.results.len());
     #[cfg(windows)]
     assert_eq!(75, entries.results.len());
     assert_eq!(0, entries.errors.len());
@@ -34,16 +34,16 @@ fn test_scandir() -> Result<(), Error> {
 }
 
 #[test]
-fn test_scandir_not_skip_hidden() -> Result<(), Error> {
+fn test_scandir_skip_hidden() -> Result<(), Error> {
     #[cfg(unix)]
     let temp_dir = common::create_temp_file_tree(3, 3, 4, 5, 6, 7)?;
     #[cfg(windows)]
     let temp_dir = common::create_temp_file_tree(3, 3, 4, 5)?;
     let scandir = Scandir::new(temp_dir.path(), Some(true))?;
-    let mut scandir = scandir.skip_hidden(false);
+    let mut scandir = scandir.skip_hidden(true);
     let entries = scandir.collect()?;
     #[cfg(unix)]
-    assert_eq!(210, entries.results.len());
+    assert_eq!(192, entries.results.len());
     #[cfg(windows)]
     assert_eq!(93, entries.results.len());
     assert_eq!(0, entries.errors.len());
@@ -73,7 +73,7 @@ fn test_scandir_extended() -> Result<(), Error> {
         .return_type(ReturnType::Ext)
         .collect()?;
     #[cfg(unix)]
-    assert_eq!(192, entries.results.len());
+    assert_eq!(210, entries.results.len());
     #[cfg(windows)]
     assert_eq!(75, entries.results.len());
     assert_eq!(0, entries.errors.len());